Upload anh

+++ backend/services/image-service/image/src/main/java/vn/uit/edu/msshop/image/domain/model/valueobject/ImageSize.java
@@ -1,5 +1,5 @@
 package vn.uit.edu.msshop.image.domain.model.valueobject;
-public record ImageSize(int value) {
+public record ImageSize(long value) {
     public ImageSize {
         if(value<=0) {
             throw new IllegalArgumentException("Invalid size");
